requirements.txt
Flask 
langchain
pathlib
python-dotenv 
requests 
slack-bolt